backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
openpyxl==3.1.2
python-docx==1.1.0
pandas==2.1.4
transformers==4.36.2
torch==2.1.2
huggingface-hub==0.20.1

